+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/atomic/CasConsensusForgettingStoreTest.java
@@ -16,18 +16,17 @@
 
 package com.palantir.atlasdb.atomic;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 import com.google.common.collect.ImmutableSet;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.keyvalue.api.Cell;
 import com.palantir.atlasdb.keyvalue.api.CheckAndSetException;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.impl.InMemoryKeyValueService;
-import org.junit.Test;
-
 import java.util.concurrent.ExecutionException;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import org.junit.Test;
 
 public class CasConsensusForgettingStoreTest {
     private static final byte[] SAD = PtBytes.toBytes("sad");
