+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/atomic/PueCassandraImitatingConsensusForgettingStoreTest.java
@@ -27,15 +27,15 @@ import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import org.junit.Test;
 
-public class CassandraImitatingConsensusForgettingStoreTest {
+public class PueCassandraImitatingConsensusForgettingStoreTest {
     private static final Cell CELL = Cell.create(new byte[] {1}, new byte[] {0});
     private static final byte[] VALUE = PtBytes.toBytes("VAL");
     private static final byte[] VALUE_2 = PtBytes.toBytes("VAL2");
     // solution to (1-x)^4 = 0.5
     private static final double PROBABILITY_THROWING_ON_QUORUM_HALF = 0.16;
-    ConsensusForgettingStore neverThrowing = new CassandraImitatingConsensusForgettingStore(0.0);
+    ConsensusForgettingStore neverThrowing = new PueCassImitatingConsensusForgettingStore(0.0);
     CassandraImitatingConsensusForgettingStore sometimesThrowing =
-            new CassandraImitatingConsensusForgettingStore(PROBABILITY_THROWING_ON_QUORUM_HALF);
+            new PueCassImitatingConsensusForgettingStore(PROBABILITY_THROWING_ON_QUORUM_HALF);
 
     @Test
     public void trivialGet() throws ExecutionException, InterruptedException {
