+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/atomic/ResilientCommitTimestampAtomicTableIntegrationTest.java
@@ -47,7 +47,7 @@ public class ResilientCommitTimestampAtomicTableIntegrationTest {
     private static final long MAXIMUM_EVALUATED_TIMESTAMP = 100;
 
     private final ConsensusForgettingStore forgettingStore =
-            new CassandraImitatingConsensusForgettingStore(WRITE_FAILURE_PROBABILITY);
+            new PueCassImitatingConsensusForgettingStore(WRITE_FAILURE_PROBABILITY);
     private final AtomicTable<Long, Long> pueTable = new ResilientCommitTimestampAtomicTable(
             forgettingStore, TwoPhaseEncodingStrategy.INSTANCE, new DefaultTaggedMetricRegistry());
